Update to only publish as web extension
