+++ .vscodeignore
@@ -1,13 +1,16 @@
 .github/**
 .vscode/**
-.vscode-test/**
-images/**
-dist/test/**
+.vscode-test-web/**
 src/**
+out/**
+images/**
+!images/icon.png
+node_modules/**
 .gitignore
 vsc-extension-quickstart.md
+webpack.config.js
+.yarnrc
 **/tsconfig.json
 **/.eslintrc.json
 **/*.map
 **/*.ts
-!images/icon.png
