+++ package.json
@@ -21,14 +21,13 @@
 		"Other"
 	],
 	"engines": {
-		"vscode": "^1.63.1"
+		"vscode": "^1.64.2"
 	},
 	"activationEvents": [
 		"onCommand:snippet-copy.copySnippet",
 		"onCommand:snippet-copy.copySnippetAsMarkdownCodeBlock"
 	],
-	"main": "./dist/extension-node.js",
-	"browser": "./dist/extension-web.js",
+	"browser": "./dist/web/extension.js",
 	"contributes": {
 		"commands": [
 			{
@@ -102,36 +101,30 @@
 		}
 	},
 	"scripts": {
-		"vscode:prepublish": "npm run package",
+		"test": "vscode-test-web --browserType=chromium --extensionDevelopmentPath=. --extensionTestsPath=dist/web/test/suite/index.js ./src/web/test/fixtures",
+		"pretest": "npm run compile-web",
+		"vscode:prepublish": "npm run package-web",
+		"compile-web": "webpack",
+		"watch-web": "webpack --watch",
+		"package-web": "webpack --mode production --devtool hidden-source-map",
 		"lint": "eslint src --ext ts",
-		"compile": "webpack",
-		"watch": "webpack --watch",
-		"pretest": "npm run lint && npm run compile",
-		"test": "npm run test-node && npm run test-web",
-		"test-node": "node ./dist/test/runTest.js",
-		"test-web": "vscode-test-web ./src/test/fixtures --browserType=chromium --extensionDevelopmentPath=. --extensionTestsPath=./dist/test/suite/index-web.js",
-		"package": "webpack --mode production --devtool hidden-source-map"
+		"run-in-browser": "vscode-test-web --browserType=chromium --extensionDevelopmentPath=. ."
 	},
 	"devDependencies": {
-		"@types/glob": "^7.2.0",
 		"@types/mocha": "^9.1.0",
-		"@types/node": "^16.11.21",
-		"@types/vscode": "^1.63.1",
+		"@types/vscode": "^1.64.0",
 		"@types/webpack-env": "^1.16.3",
-		"@typescript-eslint/eslint-plugin": "^5.10.0",
-		"@typescript-eslint/parser": "^5.10.0",
-		"@vscode/test-electron": "^2.1.0",
-		"@vscode/test-web": "0.0.18",
+		"@typescript-eslint/eslint-plugin": "^5.12.0",
+		"@typescript-eslint/parser": "^5.12.0",
+		"@vscode/test-web": "^0.0.22",
 		"assert": "^2.0.0",
-		"eslint": "^8.7.0",
-		"glob": "^7.2.0",
-		"mocha": "^9.1.4",
-		"path-browserify": "^1.0.1",
+		"eslint": "^8.9.0",
+		"mocha": "^9.2.0",
 		"process": "^0.11.10",
 		"testdouble": "^3.16.4",
 		"ts-loader": "^9.2.6",
-		"typescript": "^4.5.4",
-		"webpack": "^5.66.0",
-		"webpack-cli": "^4.9.1"
+		"typescript": "^4.5.5",
+		"webpack": "^5.68.0",
+		"webpack-cli": "^4.9.2"
 	}
 }
