--- src/test/runTest.ts
@@ -1,29 +1,0 @@
-import { runTests } from '@vscode/test-electron';
-import * as path from 'path';
-
-
-async function main() {
-	try {
-		// The folder containing the Extension Manifest package.json
-		// Passed to `--extensionDevelopmentPath`
-		const extensionDevelopmentPath = path.resolve(__dirname, '../../');
-
-		// The path to test runner
-		// Passed to --extensionTestsPath
-		const extensionTestsPath = path.resolve(__dirname, './suite/index-node');
-
-		const launchArgs = [
-			path.resolve(__dirname, './fixtures'),
-			'--disable-extensions'
-		];
-
-		// Download VS Code, unzip it and run the integration test
-		await runTests({ extensionDevelopmentPath, extensionTestsPath, launchArgs });
-	} catch (err) {
-		console.error('Failed to run tests');
-		process.exit(1);
-	}
-}
-
-// eslint-disable-next-line @typescript-eslint/no-floating-promises
-main();
