--- src/test/suite/index-web.ts
@@ -1,29 +1,0 @@
-require('mocha/mocha'); // import the mocha web build
-
-export function run(): Promise<void> {
-
-  return new Promise((c, e) => {
-    mocha.setup({
-      ui: 'bdd',
-      reporter: undefined
-    });
-
-    // bundles all files in the current directory matching `*.test`
-    const importAll = (r: __WebpackModuleApi.RequireContext) => r.keys().forEach(r);
-    importAll(require.context('.', true, /\.test$/));
-
-    try {
-      // Run the mocha test
-      mocha.run(failures => {
-        if (failures > 0) {
-          e(new Error(`${failures} tests failed.`));
-        } else {
-          c();
-        }
-      });
-    } catch (err) {
-      console.error(err);
-      e(err);
-    }
-  });
-}
