+++ tsconfig.json
@@ -4,7 +4,7 @@
 		"target": "es6",
 		"outDir": "dist",
 		"lib": [
-			"es6"
+			"es6", "WebWorker"
 		],
 		"sourceMap": true,
 		"rootDir": "src",
