+++ webpack.config.js
@@ -3,23 +3,22 @@
 
 /** @typedef {import('webpack').Configuration} WebpackConfig **/
 
-const glob = require('glob');
 const path = require('path');
 const webpack = require('webpack');
 
-/** @type {WebpackConfig} */
-const webConfig = {
-	context: __dirname,
+/** @type WebpackConfig */
+const webExtensionConfig = {
 	mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
-	target: 'webworker', // web extensions run in a webworker context
+	target: 'webworker', // extensions run in a webworker context
 	entry: {
-		'extension-web': './src/extension.ts', // source of the web extension main file
-		'test/suite/index-web': './src/test/suite/index-web.ts', // source of the web extension test runner
+		'extension': './src/web/extension.ts',
+		'test/suite/index': './src/web/test/suite/index.ts'
 	},
 	output: {
 		filename: '[name].js',
-		path: path.join(__dirname, './dist'),
+		path: path.join(__dirname, './dist/web'),
 		libraryTarget: 'commonjs',
+		devtoolModuleFilenameTemplate: '../../[resource-path]'
 	},
 	resolve: {
 		mainFields: ['browser', 'module', 'main'], // look for `browser` entry point in imported node modules
@@ -31,23 +30,17 @@ const webConfig = {
 			// Webpack 5 no longer polyfills Node.js core modules automatically.
 			// see https://webpack.js.org/configuration/resolve/#resolvefallback
 			// for the list of Node.js core module polyfills.
-			assert: require.resolve('assert'),
-			path: require.resolve("path-browserify"),
-			util: require.resolve("util"),
-		},
+			'assert': require.resolve('assert')
+		}
 	},
 	module: {
-		rules: [
-			{
-				test: /\.ts$/,
-				exclude: /node_modules/,
-				use: [
-					{
-						loader: 'ts-loader',
-					},
-				],
-			},
-		],
+		rules: [{
+			test: /\.ts$/,
+			exclude: /node_modules/,
+			use: [{
+				loader: 'ts-loader'
+			}]
+		}]
 	},
 	plugins: [
 		new webpack.ProvidePlugin({
@@ -58,52 +51,12 @@ const webConfig = {
 		vscode: 'commonjs vscode', // ignored because it doesn't exist
 	},
 	performance: {
-		hints: false,
+		hints: false
 	},
 	devtool: 'nosources-source-map', // create a source map that points to the original source file
+	infrastructureLogging: {
+		level: "log", // enables logging required for problem matchers
+	},
 };
 
-/** @type {WebpackConfig} */
-const nodeConfig = {
-	context: __dirname,
-	mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
-	target: 'node', // extensions run in a node context
-	entry: {
-		'extension-node': './src/extension.ts', // source of the node extension main file
-		'test/suite/index-node': './src/test/suite/index-node.ts', // source of the node extension test runner
-		'test/runTest': './src/test/runTest', // used to start the VS Code test runner (@vscode/test-electron)
-	},
-	output: {
-		filename: '[name].js',
-		path: path.join(__dirname, './dist'),
-		libraryTarget: 'commonjs',
-	},
-	resolve: {
-		mainFields: ['module', 'main'],
-		extensions: ['.ts', '.js'], // support ts-files and js-files
-	},
-	module: {
-		rules: [
-			{
-				test: /\.ts$/,
-				exclude: /node_modules/,
-				use: [
-					{
-						loader: 'ts-loader',
-					},
-				],
-			},
-		],
-	},
-	externals: {
-		vscode: 'commonjs vscode', // ignored because it doesn't exist
-		mocha: 'commonjs mocha', // don't bundle
-		'@vscode/test-electron': 'commonjs @vscode/test-electron' // don't bundle
-	},
-	performance: {
-		hints: false,
-	},
-	devtool: 'nosources-source-map', // create a source map that points to the original source file
-};
-
-module.exports = [webConfig, nodeConfig];
+module.exports = [webExtensionConfig];
